requirements.txt
python-dotenv==1.0.1
requests==2.31.0
Jinja2==3.1.3
